fix(segmentos): fonte unica de segmentos + quantidade fracionada (kg quebrado)
DOIS problemas, com a mesma raiz: listas paralelas que ninguem cobrava.

--- 1. O onboarding recusava "Serigrafia" ---

Mesmo com o segmento existindo no registry, no backend, nos cards e no tipo do
frontend, cadastrar a loja falhava com:

  Invalid enum value. Expected 'assistencia_tecnica' | 'oficina_mecanica' |
  'mercado' | ... , received 'serigrafia'

Faltava a QUINTA lista: o `z.enum` de sign-in.schema.ts. Ela era um `as const`
solto, sem vinculo com `BusinessSegment`, entao o vue-tsc nao tinha o que
cobrar -- o erro so aparecia na cara do usuario, no meio do cadastro.

Agora ha uma fonte unica (shared/constants/segmentos.ts) e tudo deriva dela: o
tipo, o enum do Zod, o `Segmento` do useSegmento, os cards e as dicas.

E a lista deixou de ser cobravel apenas por disciplina: CARDS e SEGMENT_TIPS
viraram `Record<Segmento, ...>`, entao segmento novo NAO COMPILA ate ganhar
card e dica. Verificado na pratica -- com um 'grafica' na lista, o build para
com dois TS2741 apontando exatamente o que falta.

Sobra um espelho que o TypeScript nao alcanca: SEGMENTOS_VALIDOS em
app/schemas/auth.py. Esta anotado no topo do arquivo novo.

--- 2. Quilo quebrado (2,5 kg) ---

Decisao do dono: a quantidade em quilo e escolha do cliente no pedido, entao
precisa ser fracionada.

E NAO HOUVE MIGRATION, de proposito. O SQLite guarda 2.5 numa coluna declarada
INTEGER: a afinidade so converte quando e sem perda, entao o valor vira `real`
e SUM/produto saem corretos -- verificado. Um `batch_alter_table` recriaria a
tabela inteira no banco vivo de duas lojas em producao, sem backup automatico
antes, para obter um comportamento que o banco JA tem.

Mudou o tipo em estoque (3 colunas), movimentacao_estoque (3) e item de OS (1),
os schemas correspondentes, e as 3 multiplicacoes de valor do item passaram a
arredondar -- valor_total e inteiro em centavos, e 2.5 * 3000 da 7500.0.

Vendas e Orcamentos seguem inteiros: a sacola sai por OS, e alargar dali sem
necessidade mexeria em codigo que as duas lojas usam todo dia.

370 pytest, vue-tsc 0.

+++ backend-fastapi/app/db/models/movimentacao_estoque.py
@@ -4,7 +4,7 @@
 #            Registra todo histórico de entradas, saídas e ajustes de estoque.
 # ---------------------------------------------------------------------------
 
-from sqlalchemy import Integer, String, Text, DateTime, ForeignKey, func
+from sqlalchemy import Float, Integer, String, Text, DateTime, ForeignKey, func
 from sqlalchemy.orm import relationship, Mapped, mapped_column
 from sqlalchemy.sql.sqltypes import Enum as SQLAlchemyEnum
 from datetime import datetime
@@ -71,18 +71,20 @@ class MovimentacaoEstoque(Base):
         nullable=False,
         doc="Tipo da movimentação: ENTRADA, SAIDA ou AJUSTE"
     )
-    quantidade: Mapped[int] = mapped_column(
-        Integer,
+    # Fracionadas para unidades de peso (2,5 kg). Ver a nota em db/models/estoque.py
+    # sobre por que a mudança de tipo não exigiu migration.
+    quantidade: Mapped[float] = mapped_column(
+        Float,
         nullable=False,
         doc="Quantidade movimentada (sempre positivo)"
     )
-    quantidade_anterior: Mapped[int] = mapped_column(
-        Integer,
+    quantidade_anterior: Mapped[float] = mapped_column(
+        Float,
         nullable=False,
         doc="Quantidade em estoque antes da movimentação"
     )
-    quantidade_posterior: Mapped[int] = mapped_column(
-        Integer,
+    quantidade_posterior: Mapped[float] = mapped_column(
+        Float,
         nullable=False,
         doc="Quantidade em estoque após a movimentação"
     )
